please finally no more lint and format warnings on frontend
