+++ frontend/lyric-sync-app/src/contexts/AuthContext.tsx
@@ -1,22 +1,6 @@
-import React, { createContext, useState, useEffect, useContext } from 'react';
+import React, { useState, useEffect } from 'react';
+import { User, AuthContext } from './authTypes';
 
-// Define User type in the same file for now
-export interface User {
-  id: string;
-  name: string;
-  email: string;
-  picture: string;
-  idToken: string;
-}
-
-interface AuthContextType {
-  user: User | null;
-  setUser: (_user: User | null) => void;
-}
-
-const AuthContext = createContext<AuthContextType | undefined>(undefined);
-
-// Prefix unused 'user' with underscore
 export const AuthProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   const [user, setUser] = useState<User | null>(() => {
     // Try to get user data from localStorage on initial load
@@ -38,12 +22,4 @@ export const AuthProvider: React.FC<{ children: React.ReactNode }> = ({ children
       {children}
     </AuthContext.Provider>
   );
-};
-
-export const useAuth = () => {
-  const context = useContext(AuthContext);
-  if (context === undefined) {
-    throw new Error('useAuth must be used within an AuthProvider');
-  }
-  return context;
 };
